sparkler-sce/webui/requirements.txt
aniso8601==8.0.0
astroid==2.3.3
attrs==19.3.0
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
DateTime==4.3
Flask==1.1.2
Flask-Cors==3.0.8
flask-restplus==0.13.0
future==0.18.2
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.3
pyArango==1.3.4
pylint==2.4.4
pyrsistent==0.16.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
soupsieve==2.0
typed-ast==1.4.1
urllib3==1.25.9
waitress==1.4.3
Werkzeug==0.16.1
wrapt==1.11.2
zipp==3.1.0
zope.interface==5.1.0
